Beatify: ⌘ takes the ends of a selection off the beat grid
A cut is not always where the beat is. The pickup starts a quarter of a
beat early; the tail you want to lose hangs a fraction past four. Until
now the selection snapped outward to whole beats and there was no way to
say otherwise.

So ⌘ (ctrl elsewhere) now frees a selection drag from the grid —
sweeping, dragging either end, sliding — read live, so a cut can be
swept to the beat and then eased off it without letting go. What that
buys is an OFFSET, not a fractional clip: the count of beats becomes a
float ("8.5 beats", and no group line, which is how the readout says
this one is off the grid), but dropping it into the editor still lands
it on a column and still covers whole columns. The fraction rides along
in the run's source beat, which is now an f64 end to end; clips saved
with whole beats read back unchanged.

+++ app/src-tauri/src/beatify_clip.rs
@@ -102,7 +102,10 @@ pub struct ClipPlacement {
     pub col: usize,
     pub beats: usize,
     pub source: ClipSourceRef,
-    pub source_beat: usize,
+    /// Where in the source this run starts, in beats — FRACTIONAL, since
+    /// a ⌘-freed selection can cut into a beat. Whole beats (every clip
+    /// saved before that existed) read back unchanged.
+    pub source_beat: f64,
 }
 
 /// A clip, saved or still being dragged around.
